d3d12: use uint64_t for fence values

+++ src/gallium/drivers/d3d12/d3d12_query.cpp
@@ -40,7 +40,7 @@ struct d3d12_query {
 
    pipe_resource *buffer;
    unsigned buffer_offset;
-   int fence_value;
+   uint64_t fence_value;
 
    struct list_head active_list;
    struct d3d12_resource *predicate;
